+++ src/lib/GaugeChart/index.js
@@ -9,12 +9,11 @@ import {
   arc,
   pie,
   select,
-  easeElastic,
-  scaleLinear,
-  interpolateHsl,
-  interpolateNumber,
 } from "d3";
 import PropTypes from "prop-types";
+
+import { setArcData } from "./utils";
+import { renderChart } from "./renderChart";
 
 import useDeepCompareEffect from "./customHooks";
 /*
@@ -182,7 +181,7 @@ const GaugeChart = (initialProps) => {
 
   useEffect(() => {
     const handleResize = () => {
-      var resize = true;
+      let resize = true;
 
       renderChart(
         resize,
@@ -263,299 +262,4 @@ GaugeChart.propTypes = {
   needleScale: PropTypes.number,
 };
 
-// This function update arc's datas when component is mounting or when one of arc's props is updated
-const setArcData = (props, nbArcsToDisplay, colorArray, arcData) => {
-  // We have to make a decision about number of arcs to display
-  // If arcsLength is setted, we choose arcsLength length instead of nrOfLevels
-  nbArcsToDisplay.current = props.arcsLength
-    ? props.arcsLength.length
-    : props.nrOfLevels;
 
-  //Check if the number of colors equals the number of levels
-  //Otherwise make an interpolation
-  if (nbArcsToDisplay.current === props.colors.length) {
-    colorArray.current = props.colors;
-  } else {
-    colorArray.current = getColors(props, nbArcsToDisplay);
-  }
-  //The data that is used to create the arc
-  // Each arc could have hiw own value width arcsLength prop
-  arcData.current = [];
-  for (var i = 0; i < nbArcsToDisplay.current; i++) {
-    var arcDatum = {
-      value:
-        props.arcsLength && props.arcsLength.length > i
-          ? props.arcsLength[i]
-          : 1,
-      color: colorArray.current[i],
-    };
-    arcData.current.push(arcDatum);
-  }
-};
-
-//Renders the chart, should be called every time the window is resized
-const renderChart = (
-  resize,
-  prevProps,
-  width,
-  margin,
-  height,
-  outerRadius,
-  g,
-  doughnut,
-  arcChart,
-  needle,
-  pieChart,
-  svg,
-  props,
-  container,
-  arcData
-) => {
-  updateDimensions(props, container, margin, width, height);
-  //Set dimensions of svg element and translations
-  svg.current
-    .attr("width", width.current + margin.current.left + margin.current.right)
-    .attr(
-      "height",
-      height.current + margin.current.top + margin.current.bottom
-    );
-  g.current.attr(
-    "transform",
-    "translate(" + margin.current.left + ", " + margin.current.top + ")"
-  );
-  //Set the radius to lesser of width or height and remove the margins
-  //Calculate the new radius
-  calculateRadius(width, height, outerRadius, margin, g);
-  doughnut.current.attr(
-    "transform",
-    "translate(" + outerRadius.current + ", " + outerRadius.current + ")"
-  );
-  //Setup the arc
-  arcChart.current
-    .outerRadius(outerRadius.current)
-    .innerRadius(outerRadius.current * (1 - props.arcWidth))
-    .cornerRadius(props.cornerRadius)
-    .padAngle(props.arcPadding);
-  //Remove the old stuff
-  doughnut.current.selectAll(".arc").remove();
-  needle.current.selectAll("*").remove();
-  g.current.selectAll(".text-group").remove();
-  //Draw the arc
-  var arcPaths = doughnut.current
-    .selectAll(".arc")
-    .data(pieChart.current(arcData.current))
-    .enter()
-    .append("g")
-    .attr("class", "arc");
-  arcPaths
-    .append("path")
-    .attr("d", arcChart.current)
-    .style("fill", function (d) {
-      return d.data.color;
-    });
-
-  drawNeedle(
-    resize,
-    prevProps,
-    props,
-    width,
-    needle,
-    container,
-    outerRadius,
-    g
-  );
-  //Translate the needle starting point to the middle of the arc
-  needle.current.attr(
-    "transform",
-    "translate(" + outerRadius.current + ", " + outerRadius.current + ")"
-  );
-};
-
-//Depending on the number of levels in the chart
-//This function returns the same number of colors
-const getColors = (props, nbArcsToDisplay) => {
-  const { colors } = props;
-  var colorScale = scaleLinear()
-    .domain([1, nbArcsToDisplay.current])
-    .range([colors[0], colors[colors.length - 1]]) //Use the first and the last color as range
-    .interpolate(interpolateHsl);
-  var colorArray = [];
-  for (var i = 1; i <= nbArcsToDisplay.current; i++) {
-    colorArray.push(colorScale(i));
-  }
-  return colorArray;
-};
-
-//If 'resize' is true then the animation does not play
-const drawNeedle = (
-  resize,
-  prevProps,
-  props,
-  width,
-  needle,
-  container,
-  outerRadius,
-  g
-) => {
-  const {
-    percent,
-    needleColor,
-    needleBaseColor,
-    hideText,
-    animate,
-    needleScale,
-    textComponent,
-  } = props;
-  var needleRadius = 15 * (width.current / 500), // Make the needle radius responsive
-    centerPoint = [0, -needleRadius / 2];
-  //Draw the triangle
-  //var pathStr = `M ${leftPoint[0]} ${leftPoint[1]} L ${topPoint[0]} ${topPoint[1]} L ${rightPoint[0]} ${rightPoint[1]}`;
-  const prevPercent = prevProps ? prevProps.percent : 0;
-  var pathStr = calculateRotation(
-    prevPercent || percent,
-    outerRadius,
-    width,
-    needleScale
-  );
-  needle.current.append("path").attr("d", pathStr).attr("fill", needleColor);
-  //Add a circle at the bottom of needle
-  needle.current
-    .append("circle")
-    .attr("cx", centerPoint[0])
-    .attr("cy", centerPoint[1])
-    .attr("r", needleRadius)
-    .attr("fill", needleBaseColor);
-  if (!hideText && !textComponent) {
-    addText(percent, props, outerRadius, width, g);
-  }
-  //Rotate the needle
-  if (!resize && animate) {
-    needle.current
-      .transition()
-      .delay(props.animDelay)
-      .ease(easeElastic)
-      .duration(props.animateDuration)
-      .tween("progress", function () {
-        const currentPercent = interpolateNumber(prevPercent, percent);
-        return function (percentOfPercent) {
-          const progress = currentPercent(percentOfPercent);
-          return container.current
-            .select(`.needle path`)
-            .attr(
-              "d",
-              calculateRotation(progress, outerRadius, width, needleScale)
-            );
-        };
-      });
-  } else {
-    container.current
-      .select(`.needle path`)
-      .attr("d", calculateRotation(percent, outerRadius, width, needleScale));
-  }
-};
-
-const calculateRotation = (percent, outerRadius, width, needleScale) => {
-  var needleLength = outerRadius.current * needleScale, //TODO: Maybe it should be specified as a percentage of the arc radius?
-    needleRadius = 15 * (width.current / 500),
-    theta = percentToRad(percent),
-    centerPoint = [0, -needleRadius / 2],
-    topPoint = [
-      centerPoint[0] - needleLength * Math.cos(theta),
-      centerPoint[1] - needleLength * Math.sin(theta),
-    ],
-    leftPoint = [
-      centerPoint[0] - needleRadius * Math.cos(theta - Math.PI / 2),
-      centerPoint[1] - needleRadius * Math.sin(theta - Math.PI / 2),
-    ],
-    rightPoint = [
-      centerPoint[0] - needleRadius * Math.cos(theta + Math.PI / 2),
-      centerPoint[1] - needleRadius * Math.sin(theta + Math.PI / 2),
-    ];
-  var pathStr = `M ${leftPoint[0]} ${leftPoint[1]} L ${topPoint[0]} ${topPoint[1]} L ${rightPoint[0]} ${rightPoint[1]}`;
-  return pathStr;
-};
-
-//Returns the angle (in rad) for the given 'percent' value where percent = 1 means 100% and is 180 degree angle
-const percentToRad = (percent) => {
-  return percent * Math.PI;
-};
-
-//Adds text undeneath the graft to display which percentage is the current one
-const addText = (percentage, props, outerRadius, width, g) => {
-  const { formatTextValue, fontSize } = props;
-  var textPadding = 20;
-  const text = formatTextValue
-    ? formatTextValue(floatingNumber(percentage))
-    : floatingNumber(percentage) + "%";
-  g.current
-    .append("g")
-    .attr("class", "text-group")
-    .attr(
-      "transform",
-      `translate(${outerRadius.current}, ${
-        outerRadius.current / 2 + textPadding
-      })`
-    )
-    .append("text")
-    .text(text)
-    // this computation avoid text overflow. When formatted value is over 10 characters, we should reduce font size
-    .style("font-size", () =>
-      fontSize
-        ? fontSize
-        : `${width.current / 11 / (text.length > 10 ? text.length / 10 : 1)}px`
-    )
-    .style("fill", props.textColor)
-    .style("text-anchor", "middle");
-};
-
-const floatingNumber = (value, maxDigits = 2) => {
-  return Math.round(value * 100 * 10 ** maxDigits) / 10 ** maxDigits;
-};
-
-const calculateRadius = (width, height, outerRadius, margin, g) => {
-  //The radius needs to be constrained by the containing div
-  //Since it is a half circle we are dealing with the height of the div
-  //Only needs to be half of the width, because the width needs to be 2 * radius
-  //For the whole arc to fit
-
-  //First check if it is the width or the height that is the "limiting" dimension
-  if (width.current < 2 * height.current) {
-    //Then the width limits the size of the chart
-    //Set the radius to the width - the horizontal margins
-    outerRadius.current =
-      (width.current - margin.current.left - margin.current.right) / 2;
-  } else {
-    outerRadius.current =
-      height.current - margin.current.top - margin.current.bottom;
-  }
-  centerGraph(width, g, outerRadius, margin);
-};
-
-//Calculates new margins to make the graph centered
-const centerGraph = (width, g, outerRadius, margin) => {
-  margin.current.left =
-    width.current / 2 - outerRadius.current + margin.current.right;
-  g.current.attr(
-    "transform",
-    "translate(" + margin.current.left + ", " + margin.current.top + ")"
-  );
-};
-
-const updateDimensions = (props, container, margin, width, height) => {
-  //TODO: Fix so that the container is included in the component
-  const { marginInPercent } = props;
-  var divDimensions = container.current.node().getBoundingClientRect(),
-    divWidth = divDimensions.width,
-    divHeight = divDimensions.height;
-
-  //Set the new width and horizontal margins
-  margin.current.left = divWidth * marginInPercent;
-  margin.current.right = divWidth * marginInPercent;
-  width.current = divWidth - margin.current.left - margin.current.right;
-
-  margin.current.top = divHeight * marginInPercent;
-  margin.current.bottom = divHeight * marginInPercent;
-  height.current =
-    width.current / 2 - margin.current.top - margin.current.bottom;
-  //height.current = divHeight - margin.current.top - margin.current.bottom;
-};
